+++ ali/common/src/main/java/com/yanny/ali/api/IServerRegistry.java
@@ -4,6 +4,7 @@ import com.yanny.aci.api.ICoreServerRegistry;
 import com.yanny.aci.api.RangeValue;
 import com.yanny.aci.tooltip.TooltipBuilder;
 import com.yanny.aci.tooltip.TooltipNode;
+import com.yanny.ali.plugin.glm.IDestinationResolver;
 import com.yanny.ali.plugin.server.EnchantedRanges;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.minecraft.resources.ResourceLocation;
@@ -46,6 +47,8 @@ public interface IServerRegistry extends ICoreServerRegistry<IServerUtils> {
     <T extends LootItemCondition> void registerChanceModifier(Class<T> type, TriConsumer<IServerUtils, T, EnchantedRanges> consumer);
 
     <T extends LootItemFunction> void registerItemStackModifier(Class<T> type, TriFunction<IServerUtils, T, ItemStack, ItemStack> consumer);
+
+    <T extends LootItemCondition> void registerDestination(Class<T> type, IDestinationResolver<T> resolver);
 
     void registerLootModifiers(Function<IServerUtils, List<ILootModifier<?>>> getter);