+++ ali/forge/src/main/java/com/yanny/ali/forge/plugin/ForgePlugin.java
@@ -18,10 +18,10 @@ import com.yanny.ali.forge.mixin.MixinLootTableIdCondition;
 import com.yanny.ali.language.Lang;
 import com.yanny.ali.platform.Services;
 import com.yanny.ali.plugin.common.trades.ItemsToItemsNode;
+import com.yanny.ali.plugin.glm.Destination;
 import com.yanny.ali.plugin.glm.GlobalLootModifierUtils;
 import com.yanny.ali.plugin.glm.IGlobalLootModifierPlugin;
 import com.yanny.ali.plugin.glm.IGlobalLootModifierWrapper;
-import com.yanny.ali.plugin.glm.ILootTableIdConditionPredicate;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import net.minecraftforge.common.BasicItemListing;
@@ -59,6 +59,8 @@ public class ForgePlugin implements IPlugin {
         registry.registerIngredientTooltip(PartialNBTIngredient.class, ForgeIngredientTooltipUtils::getPartialNbtIngredientTooltip);
         registry.registerIngredientTooltip(StrictNBTIngredient.class, ForgeIngredientTooltipUtils::getStrictNbtIngredientTooltip);
 
+        registry.registerDestination(LootTableIdCondition.class, ForgePlugin::getLootTableIdDestination);
+
         registry.registerItemListing(BasicItemListing.class, ForgePlugin::getBasicItemListingNode);
 
         registry.registerLootModifiers(ForgePlugin::registerLootModifiers);
@@ -94,16 +96,20 @@ public class ForgePlugin implements IPlugin {
     }
 
     @NotNull
+    public static Destination getLootTableIdDestination(IServerUtils ignoredUtils, LootTableIdCondition cond) {
+        return new Destination.Table(((MixinLootTableIdCondition) cond).getTargetLootTableId(), true);
+    }
+
+    @NotNull
     private static List<ILootModifier<?>> registerLootModifiers(IServerUtils utils) {
         Map<Class<?>, BiFunction<IServerUtils, IGlobalLootModifier, Optional<ILootModifier<?>>>> glmMap = new HashMap<>();
         Set<Class<?>> missingGLM = new HashSet<>();
         List<ILootModifier<?>> lootModifiers = new ArrayList<>();
-        ILootTableIdConditionPredicate tablePredicate = getLootTableIdConditionPredicate();
         IGlobalLootModifierPlugin.IRegistry forgeRegistry = getForgeRegistry(glmMap);
 
         for (IPlugin plugin : Services.getPlatform().getPlugins()) {
             if (plugin instanceof IGlobalLootModifierPlugin forgePlugin) {
-                forgePlugin.registerGlobalLootModifier(forgeRegistry, tablePredicate);
+                forgePlugin.registerGlobalLootModifier(forgeRegistry);
             }
         }
 
@@ -124,7 +130,7 @@ public class ForgePlugin implements IPlugin {
                         LOGGER.warn("Unable to locate destination for GLM {}", wrapper.getName());
                     }
                 } else {
-                    Optional<ILootModifier<?>> modifier = GlobalLootModifierUtils.getMissingGlobalLootModifier(utils, wrapper, tablePredicate);
+                    Optional<ILootModifier<?>> modifier = GlobalLootModifierUtils.getMissingGlobalLootModifier(utils, wrapper);
 
                     missingGLM.add(globalLootModifier.getClass());
 
@@ -142,21 +148,6 @@ public class ForgePlugin implements IPlugin {
         missingGLM.forEach((c) -> LOGGER.warn("Missing GLM for {}", c.getName()));
 
         return lootModifiers;
-    }
-
-    @NotNull
-    private static ILootTableIdConditionPredicate getLootTableIdConditionPredicate() {
-        return new ILootTableIdConditionPredicate() {
-            @Override
-            public boolean isLootTableIdCondition(LootItemCondition condition) {
-                return condition instanceof LootTableIdCondition;
-            }
-
-            @Override
-            public ResourceLocation getTargetLootTableId(LootItemCondition condition) {
-                return ((MixinLootTableIdCondition) condition).getTargetLootTableId();
-            }
-        };
     }
 
     @NotNull